Split order create types and cleanup code
