--- src/connect/ValidationError.ts
@@ -1,1 +1,0 @@
-export class ValidationError extends Error {}
