+++ src/connect/index.ts
@@ -3,4 +3,3 @@ export * from './ApiClient';
 export * from './PayNLProvider';
 export * from './order/Order';
 export * from './ApiError';
-export * from './ValidationError';
