+++ src/connect/order/Amount.ts
@@ -1,0 +1,9 @@
+export type CreateAmount = {
+    value: number;
+    currency?: string;
+};
+
+export type ResponseAmount = {
+    value: number;
+    currency: string;
+};
