+++ src/connect/order/Customer.ts
@@ -1,0 +1,21 @@
+export type Customer = {
+    email?: string;
+    firstName?: string;
+    lastName?: string;
+    birthDate?: Date | string;
+    gender?: string;
+    phone?: string;
+    language?: string;
+    /** Locale code like "en_US" or "nl_BE" */
+    locale?: string;
+    ipAddress?: string;
+    trust?: number;
+    reference?: string;
+    gaClientId?: string;
+    company?: {
+        name?: string;
+        cocNumber?: string;
+        vatNumber?: string;
+        country?: string;
+    };
+};
