+++ src/connect/order/Integration.ts
@@ -1,0 +1,4 @@
+export type Integration = {
+    test?: boolean;
+    pointOfInteraction?: string;
+};
