+++ src/connect/order/Notification.ts
@@ -1,0 +1,4 @@
+export type Notification = {
+    type?: 'push' | 'email';
+    recipient?: string;
+};
