+++ src/connect/order/Stats.ts
@@ -1,0 +1,10 @@
+export type Stats = {
+    tool?: string;
+    info?: string;
+    object?: string;
+    extra1?: string;
+    extra2?: string;
+    extra3?: string;
+    domainId?: string;
+    promotorId?: number;
+};
