+++ tests/connect/ApiError.test.ts
@@ -15,8 +15,8 @@ const BadRequestResponse = {
 } as Response;
 
 describe('ApiError', () => {
-    it('should include HTTP status code and message', () => {
-        const subject = new ApiError(
+    it('should include HTTP status code and message', async () => {
+        const subject = await ApiError.create(
             new ApiResponse(new Response(null, { status: 418, statusText: "I'm a teapot" })),
         );
 
@@ -28,7 +28,7 @@ describe('ApiError', () => {
     it('should have the Pay.nl API error response', async () => {
         const expectedError = BadRequestBody;
 
-        const subject = new ApiError(new ApiResponse(BadRequestResponse));
+        const subject = await ApiError.create(new ApiResponse(BadRequestResponse));
 
         expect(await subject.body()).toEqual(expectedError);
     });
@@ -50,13 +50,13 @@ describe('ApiError', () => {
             },
         } as Response;
 
-        const subject = new ApiError(new ApiResponse(mockResponse));
+        const subject = await ApiError.create(new ApiResponse(mockResponse));
 
         expect(await subject.body()).toEqual(expectedError);
     });
 
     it('should allow to fetch error body twice', async () => {
-        const subject = new ApiError(new ApiResponse(BadRequestResponse));
+        const subject = await ApiError.create(new ApiResponse(BadRequestResponse));
 
         const firstTime = await subject.body();
         const secondTime = await subject.body();
