+++ tests/connect/order/OrderApi.test.ts
@@ -3,7 +3,7 @@ import { ApiClient } from '../../../src';
 import { FetchMock } from '../support/FetchMock';
 import { mockClientOptions } from '../support/mockClientOptions';
 
-describe('OrdersApi', () => {
+describe('OrderApi', () => {
     it('can send order:create', async () => {
         const clientMock = new ApiClient(mockClientOptions);
 
@@ -24,21 +24,5 @@ describe('OrdersApi', () => {
         });
 
         expect(response.id).toBe('68ac5742-4223-812b-1225-4060087323de');
-    });
-
-    it('cannot send order:create with invalid reference', async () => {
-        const clientMock = new ApiClient(mockClientOptions);
-
-        const subject = new OrderApi(clientMock);
-
-        await expect(async () => {
-            await subject.create({
-                reference: 'ABC-123',
-                amount: {
-                    value: 1000,
-                    currency: 'EUR',
-                },
-            });
-        }).rejects.toThrow('The order reference may only contain alphanumeric characters.');
     });
 });
